Relocate unit tests
