+++ test/common/schemaform/definitions/applicantInformation.unit.spec.jsx
@@ -4,8 +4,8 @@ import { expect } from 'chai';
 import sinon from 'sinon';
 import ReactTestUtils from 'react-dom/test-utils';
 
-import { DefinitionTester, submitForm } from '../../util/schemaform-utils.jsx';
-import applicantInformation from '../../../src/js/edu-benefits/pages/applicantInformation.js';
+import { DefinitionTester, submitForm } from '../../../util/schemaform-utils.jsx';
+import applicantInformation from '../../../../src/js/common/schemaform/definitions/applicantInformation.js';
 
 import fullSchema1990e from 'vets-json-schema/dist/22-1990E-schema.json';
