+++ test/common/schemaform/definitions/personId.unit.spec.jsx
@@ -4,8 +4,8 @@ import { expect } from 'chai';
 import sinon from 'sinon';
 import ReactTestUtils from 'react-dom/test-utils';
 
-import { DefinitionTester, submitForm } from '../../util/schemaform-utils.jsx';
-import * as personId from '../../../src/common/schemaform/definitions/personId';
+import { DefinitionTester, submitForm } from '../../../util/schemaform-utils.jsx';
+import * as personId from '../../../../src/js/common/schemaform/definitions/personId';
 import commonDefinitions from 'vets-json-schema/dist/definitions.json';
 
 describe('Edu personId', () => {
